recipes-tisdk: edgeai-components: tiovx-apps: Add three patches
Add two patches for the Phytec phycam VM-016 (similar to the patches for
edgeai-tiovx-modules, as we have a 'copy' of modules in here).

Add a patch to fix overflow due to very long model artifact path.

While at it, remove the anyway unused patch from a previous SDK version
and fixed commit message typo of another.

Signed-off-by: Steffen Hemer <s.hemer@phytec.de>