recipes-tisdk: ti-psdk-rtos: ti-vision-apps: Fix repo/git issues
There have been multiple issues in the past: git.ti.com removed git
protocol but manifest entries for Kirkstone-compatible release was not
adopted accordingly. The processor-sdk/imaging repository is missing the
'REL.PSDK.ANALYTICS.09.02.00.05' tag to work with the manifest and last
but not least the repo fetcher does not work in combination with the
default DL_DIR and the PHYTEC way to init the metalayer (release) states
also using repo (alternate DL_DIR outside the Yocto tree structure does
work though).

So instead of using the manifest, add the list of relevant git repositories
directly to the SRC_URI.

Signed-off-by: Steffen Hemer <s.hemer@phytec.de>